+++ backend/canisters/online_users/impl/src/lifecycle/init.rs
@@ -13,7 +13,12 @@ fn init(args: Args) {
     init_cycles_dispenser_client(args.cycles_dispenser_canister_id, args.test_mode);
 
     let env = init_env([0; 32]);
-    let data = Data::new(args.user_index_canister_id, args.cycles_dispenser_canister_id, args.test_mode);
+    let data = Data::new(
+        args.user_index_canister_id,
+        args.event_relay_canister_id,
+        args.cycles_dispenser_canister_id,
+        args.test_mode,
+    );
 
     init_state(env, data, args.wasm_version);