+++ frontend/openchat-agent/src/services/user/candid/idl.js
@@ -161,8 +161,9 @@ export const idlFactory = ({ IDL }) => {
   });
   const Rules = IDL.Record({ 'text' : IDL.Text, 'enabled' : IDL.Bool });
   const VerifiedCredentialGate = IDL.Record({
-    'credential' : IDL.Text,
-    'issuer' : IDL.Text,
+    'credential_arguments' : IDL.Opt(IDL.Vec(IDL.Nat8)),
+    'issuer_origin' : IDL.Text,
+    'credential_type' : IDL.Text,
   });
   const SnsNeuronGate = IDL.Record({
     'min_stake_e8s' : IDL.Opt(IDL.Nat64),